requirements.txt
pytest==7.0.1
allure-pytest==2.9.45
allure-python-commons==2.9.45
pytest-xdist==2.5.0
webdriver-manager==3.5.3
selenium==4.1.3


